Refactoring. Custom views now extend a ViewCustom to make code simpler and more consistant. Storage of current tags is a bit neater.

+++ SimpleRss/src/main/java/com/poloure/simplerss/AdapterManageFragments.java
@@ -46,7 +46,7 @@ class AdapterManageFragments extends BaseAdapter
    public
    long getItemId(int position)
    {
-      return (long) position;
+      return position;
    }
 
    @Override
